[TS-244] Rename: CardDetail -> StarCardDetail로 이름 변경
CardDetail이라고 하면 어떤 카드인지 명확하지 않아서 구체적인 이름으로 변경
