+++ src/components/StarCardDetail/Button.tsx
@@ -1,6 +1,6 @@
 // TODO: Button 공통 컴포넌트 만들어지면 교체 예정
 
-import { buttonStyle } from "@/components/CardDetail/Button.style";
+import { buttonStyle } from "@/components/StarCardDetail/Button.style";
 
 interface ButtonProps {
 	children: React.ReactNode;
