+++ src/components/StarCardDetail/CategoryLabel/CategoryLabel.style.ts
