+++ src/components/StarCardDetail/CategoryLabel/CategoryLabel.tsx
@@ -1,5 +1,5 @@
-import { categoryLabelStyle } from "@/components/CardDetail/CategoryLabel/CategoryLabel.style";
-import Label from "@/components/CardDetail/CategoryLabel/Label";
+import { categoryLabelStyle } from "@/components/StarCardDetail/CategoryLabel/CategoryLabel.style";
+import Label from "@/components/StarCardDetail/CategoryLabel/Label";
 
 // TODO: API 연결 후 삭제 예정
 const data = [
