+++ src/components/StarCardDetail/CategoryLabel/Label.style.ts
