+++ src/components/StarCardDetail/CategoryLabel/Label.tsx
@@ -1,4 +1,4 @@
-import { labelStyle } from "@/components/CardDetail/CategoryLabel/Label.style";
+import { labelStyle } from "@/components/StarCardDetail/CategoryLabel/Label.style";
 
 export default function Label({ children }: { children: string }) {
 	return <div css={labelStyle}>{children}</div>;
