+++ src/components/StarCardDetail/Img/Img.style.ts
