+++ src/components/StarCardDetail/Img/PossessionLabel.tsx
@@ -1,4 +1,4 @@
-import { possessionLabelStyle } from "@/components/CardDetail/Img/PossessionLabel.style";
+import { possessionLabelStyle } from "@/components/StarCardDetail/Img/PossessionLabel.style";
 
 export default function PossessionLabel() {
 	return <div css={possessionLabelStyle}>보유</div>;
