+++ src/components/StarCardDetail/Img/Star.style.ts
