+++ src/components/StarCardDetail/Story.style.ts
