+++ src/components/StarCardDetail/Story.tsx
