+++ src/components/StarCardDetail/Title.style.ts
