+++ src/constants/starCardDetailConstants.ts
