+++ src/pages/StarCardDetailPage.style.ts
