+++ src/pages/StarCardDetailPage.tsx
@@ -1,11 +1,11 @@
-import Button from "@/components/CardDetail/Button";
-import CategoryLabel from "@/components/CardDetail/CategoryLabel/CategoryLabel";
-import Img from "@/components/CardDetail/Img/Img";
-import Story from "@/components/CardDetail/Story";
-import Title from "@/components/CardDetail/Title";
-import { buttonState } from "@/constants/cardDetailConstants";
+import Button from "@/components/StarCardDetail/Button";
+import CategoryLabel from "@/components/StarCardDetail/CategoryLabel/CategoryLabel";
+import Img from "@/components/StarCardDetail/Img/Img";
+import Story from "@/components/StarCardDetail/Story";
+import Title from "@/components/StarCardDetail/Title";
+import { buttonState } from "@/constants/starCardDetailConstants";
 
-import { sectionStyle, headerStyle, buttonContainerStyle } from "@/pages/CardDetailPage.style";
+import { sectionStyle, headerStyle, buttonContainerStyle } from "@/pages/StarCardDetailPage.style";
 
 // TODO: API 연결 후 삭제 예정(상태에 따른 Img, Button UI 변경 확인 용)
 interface dataType {
@@ -18,7 +18,7 @@ const data: dataType = {
 };
 
 // TODO: Header, Button 공통 컴포넌트 머지되면 교체 예정
-export default function CardDetailPage() {
+export default function StarCardDetailPage() {
 	return (
 		<>
 			<header css={headerStyle}></header>
